refactor: lib 디렉토리의 actions 파일 삭제
