+++ app/lib/data.ts
@@ -1,28 +1,6 @@
 import db from '@/lib/db';
 import getSession from '@/lib/session';
 
-export async function fetchBooks(userId: number) {
-  try {
-    const bookshelf = await db.bookshelf.findUnique({
-      where: {
-        ownerId: userId,
-      },
-      include: {
-        books: true,
-      },
-    });
-
-    if (!bookshelf) {
-      throw new Error(`Bookshelf not found for user(id: ${userId})`);
-    }
-
-    return bookshelf.books;
-  } catch (error) {
-    console.error('Database Error:', error);
-    throw new Error(`Failed to fetch books of user(id : ${userId})`);
-  }
-}
-
 export async function getUser() {
   const session = await getSession();
